// frontend/src/api/models/listIncomeIncomesParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { ListIncomeIncomesCategory } from './listIncomeIncomesCategory'
import type { ListIncomeIncomesReceiptStatus } from './listIncomeIncomesReceiptStatus'
import type { ListIncomeIncomesStatus } from './listIncomeIncomesStatus'

export type ListIncomeIncomesParams = {
  /**
   * Search keyword
   */
  search?: string
  /**
   * Filter by category
   */
  category?: ListIncomeIncomesCategory
  /**
   * Filter by status
   */
  status?: ListIncomeIncomesStatus
  /**
   * Filter by receipt status
   */
  receipt_status?: ListIncomeIncomesReceiptStatus
  /**
   * Filter from date (YYYY-MM-DD)
   */
  from_date?: string
  /**
   * Filter to date (YYYY-MM-DD)
   */
  to_date?: string
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   */
  page_size?: number
}