// frontend/src/api/models/listIncomeIncomesCategory.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListIncomeIncomesCategory =
  (typeof ListIncomeIncomesCategory)[keyof typeof ListIncomeIncomesCategory]

export const ListIncomeIncomesCategory = {
  INVESTMENT: 'INVESTMENT',
  SUBSIDY: 'SUBSIDY',
  INTEREST: 'INTEREST',
  RENTAL: 'RENTAL',
  REFUND: 'REFUND',
  COMPENSATION: 'COMPENSATION',
  ASSET_DISPOSAL: 'ASSET_DISPOSAL',
  OTHER: 'OTHER',
} as const