// frontend/src/api/models/listIncomeIncomesStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListIncomeIncomesStatus =
  (typeof ListIncomeIncomesStatus)[keyof typeof ListIncomeIncomesStatus]

export const ListIncomeIncomesStatus = {
  DRAFT: 'DRAFT',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
} as const