// frontend/src/api/models/listIncomeIncomesReceiptStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListIncomeIncomesReceiptStatus =
  (typeof ListIncomeIncomesReceiptStatus)[keyof typeof ListIncomeIncomesReceiptStatus]

export const ListIncomeIncomesReceiptStatus = {
  PENDING: 'PENDING',
  RECEIVED: 'RECEIVED',
} as const