// frontend/src/api/models/handlerIncomeSummaryResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerIncomeSummaryResponseByCategory } from './handlerIncomeSummaryResponseByCategory'

/**
 * Income summary response
 */
export interface HandlerIncomeSummaryResponse {
  by_category?: HandlerIncomeSummaryResponseByCategory
  total_confirmed?: number
  total_draft?: number
}